+++ src/Repository/PersistableEntityInterface.php
@@ -16,8 +16,8 @@ interface PersistableEntityInterface
     public function toElastic(): array;
 
     /**
-     * @param array $document
-     * @param array $metaData
+     * @param array $document the raw document as found in the _source field of the raw Elasticsearch response
+     * @param array $metaData contains all "underscore-fields" delivered in the raw Elasticsearch response (e.g. _score)
      *
      * @return mixed
      */