+++ tests/Repository/RepositoryConfigurationTest.php
@@ -7,6 +7,7 @@ use Kununu\Elasticsearch\Exception\RepositoryConfigurationException;
 use Kununu\Elasticsearch\Repository\EntityFactoryInterface;
 use Kununu\Elasticsearch\Repository\EntitySerializerInterface;
 use Kununu\Elasticsearch\Repository\OperationType;
+use Kununu\Elasticsearch\Repository\PersistableEntityInterface;
 use Kununu\Elasticsearch\Repository\RepositoryConfiguration;
 use PHPUnit\Framework\TestCase;
 
@@ -209,5 +210,44 @@ class RepositoryConfigurationTest extends TestCase
         );
 
         $config = new RepositoryConfiguration(['entity_factory' => $myFactory]); // NOSONAR
+    }
+
+    public function testValidEntityClass(): void
+    {
+        $myEntity = new class implements PersistableEntityInterface
+        {
+            public function toElastic(): array
+            {
+                return [];
+            }
+
+            public static function fromElasticDocument(array $document, array $metaData)
+            {
+            }
+        };
+
+        $config = new RepositoryConfiguration(['entity_class' => get_class($myEntity)]);
+
+        $this->assertEquals(get_class($myEntity), $config->getEntityClass());
+    }
+
+    public function testInvalidEntityClass(): void
+    {
+        $this->expectException(RepositoryConfigurationException::class);
+        $this->expectExceptionMessage(
+            'Invalid entity class given. Must be of type \Kununu\Elasticsearch\Repository\PersistableEntityInterface'
+        );
+
+        $config = new RepositoryConfiguration(['entity_class' => \stdClass::class]); // NOSONAR
+    }
+
+    public function testNonExistentEntityClass(): void
+    {
+        $this->expectException(RepositoryConfigurationException::class);
+        $this->expectExceptionMessage(
+            'Given entity class does not exist.'
+        );
+
+        $config = new RepositoryConfiguration(['entity_class' => '\Foo\Bar']); // NOSONAR
     }
 }